+++ packages/flow-core/src/components/f-accordion/f-accordion.ts
@@ -1,4 +1,4 @@
-import { html, unsafeCSS } from "lit";
+import { html, PropertyValueMap, unsafeCSS } from "lit";
 import { property, query } from "lit/decorators.js";
 import eleStyle from "./f-accordion.scss";
 import { FRoot } from "../../mixins/components/f-root/f-root";
@@ -164,7 +164,10 @@ export class FAccordion extends FRoot {
 			</f-div>
 		`;
 	}
-	updated() {
+	protected async updated(
+		changedProperties: PropertyValueMap<any> | Map<PropertyKey, unknown>
+	): Promise<void> {
+		super.updated(changedProperties);
 		this.handleAccordionBody();
 	}
 }