+++ packages/flow-core/CHANGELOG.md
@@ -10,6 +10,10 @@
 
 - `f-breadcrumb` : component added.
 
+### Bug Fixes
+
+- Fixed - `tooltip` wasn't working on various components as they were not inheriting properties from root component
+
 ## [1.15.4] - 2023-05-31
 
 ### Bug Fixes